refactor(utilities): Move utility libraries to directory for easier partial overrides
